requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-socketio==5.1.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
redis==4.1.0